feat: implement LTS 1.77.2 compatibility and add BuildOmsLinuxBinary function
